pyproject.toml
[project]
name = "komonbot"
version = "0.1.0"
description = "Telegram Web App + FastAPI backend + Ghost CMS integration for cultural space management"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "aiogram>=3.14.0",
    "httpx>=0.28.0",
    "apscheduler>=3.10.4",
    "slowapi>=0.1.9",
    "bleach>=6.2.0",
    "markupsafe>=3.0.0",
    "tenacity>=9.0.0",
    "structlog>=24.4.0",
    "PyJWT>=2.10.0",
    "python-multipart>=0.0.18",
    "email-validator>=2.2.0",
    "psycopg2-binary>=2.9.11",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "respx>=0.22.0",
    "ruff>=0.8.0",
    "black>=24.10.0",
    "coverage>=7.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["N802", "UP042", "UP046"]

[tool.black]
target-version = ["py312"]
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["src"]
